#50 resolved.
Removed the global lock in main.go. The user of unique ids for both sessions
and connections removes any worry of
race conditions.

Added the DoWrite method to ManageedDB. It takes functions
that do writes to the database and runs them inside a mutex that is
opaque to calling code.

Modified the db-writing code in model.go to use the new mechanism.